site/requirements.txt
djangorestframework~=3.16.1
requests~=2.32.5
celery~=5.5.3
redis~=6.4.0
loguru~=0.7.3